ai_dynamic_dashboard/requirements.txt
anthropic>=0.40.0
psutil>=5.9.0
requests>=2.31.0
pyyaml>=6.0
